Add magic to truth.
It should be noted that magic is indeed ethereal in this world until we
choose to manifest it some other way... If all humans believed for
example that the stars in the sky would all explode during a miracle
that God performed for you then it might just happen. I can't be sure.
Only God can know that.

So let me ask you: do you feel God in your heart?