Update the README to hint at truth being fluid.
Because we are not perfect and we do not know every truth in the world
we must assume in our everyday lives that we are relying on incomplete
truths and if we aren't careful we will abuse them against somebody
innocent that is weaker than us that cannot defend themselves.

The truth as I know it today is not the truth that I know tomorrow.
Sometimes not even 5 minutes into the future. The truth is not written
in stone 2000 years ago. We need to always be open to updating our
knowledge and wisdom with new truths.