Failed attempt to reimplement using package.
I'm not familiar with Text::Balanced, and it proves to be a bit more
low-level than I had hoped. So I'll either need to learn how to use it
or find a higher level package to suit my needs. That or someone else
will have to finish this branch for me. Whatever comes first.