Fix a few classes that got out of whack.
The manual management of these classes/tags certainly has flaws,
particularly having to duplicate it in both class and title attributes.
Arguably the title attribute isn't necessary, though for accessibility
reasons I am hesitant to remove it and rely instead on just JavaScript
to render the classes to the user.

It occurred to me that perhaps instead of the author having to always
annotate everything it would make better sense if their editor(s) and/or
devices/software would be smart enough to track your usage of a given
language and phrases within it and what you typically mean by them so
that before long the computer can fill in most of the between the lines
just by knowing your writing style... And then it could also help to
highlight new ones or any oddities that it can't explain yet to help you
spot mistakes and to help the computer learn from changes.