Add experimental scramble and unscramble buttons.
The idea being to try to tell a parallel story without changing the
meaning, but pehraps by avoiding certain triggering words or classes. Of
course, without any analytics or configuration to tell us which ones to
avoid, this is just pseudo-random.

And since the classes that I've used are a bit "extra" it doesn't work
very well to transform into a document that you can read.