Update the truth document with my second video.
If it's not evident yet this is a live document that will not
necessarily remain static. It could change over time to represent
whatever the god of its host wishes. In other words, whatever I perceive
to be the truth or the way to the truth at any given time.